In migration, create one label per channel using UID instead of name
Previously, we used the notification channel names to create the routing labels
on an alert rule. This makes future work to individually re-migrate
single channels difficult as channel names can be easily changed and are not
guaranteed unique by any db constraint.

This change modifies the logic to instead use the channel uid, which is
guaranteed unique per org at the db level and is significantly less likely to
change (indeed even if it does change, we can likely assume that the new channel
 is not equal in identity to the old one).

In addition, we move away from a single `__contact__` label with an array of
receiver names combined with regex-based route matching to multiple
`__contact_{uid}__` labels and a simple equality-based route matching. These
routes are now all nested under a single top-level route matched against
`__use_legacy_channels__ = true` for ease of organization.

This is done to improve the experience for users wanting to keep the
default migrated routing strategy but modify which contact points an alert
sends to. Editing the array could be difficult when large.

This has the added benefit of great simplifying the logic around contact point
migration as well and removes the (now unnecessary) DashAlert wrapper for
 legacymodels.Alert.